rustc_codegen_ssa: move local variable debuginfo to mir::debuginfo.
